+++ src/main/kotlin/com/github/ajalt/colormath/CssColors.kt
@@ -1,0 +1,311 @@
+package com.github.ajalt.colormath
+
+/**
+ * All css color keywords.
+ */
+object CssColors {
+    val black = RGB(0, 0, 0)
+    val silver = RGB(192, 192, 192)
+    val gray = RGB(128, 128, 128)
+    val white = RGB(255, 255, 255)
+    val maroon = RGB(128, 0, 0)
+    val red = RGB(255, 0, 0)
+    val purple = RGB(128, 0, 128)
+    val fuchsia = RGB(255, 0, 255)
+    val green = RGB(0, 128, 0)
+    val lime = RGB(0, 255, 0)
+    val olive = RGB(128, 128, 0)
+    val yellow = RGB(255, 255, 0)
+    val navy = RGB(0, 0, 128)
+    val blue = RGB(0, 0, 255)
+    val teal = RGB(0, 128, 128)
+    val aqua = RGB(0, 255, 255)
+    val orange = RGB(255, 165, 0)
+    val aliceblue = RGB(240, 248, 255)
+    val antiquewhite = RGB(250, 235, 215)
+    val aquamarine = RGB(127, 255, 212)
+    val azure = RGB(240, 255, 255)
+    val beige = RGB(245, 245, 220)
+    val bisque = RGB(255, 228, 196)
+    val blanchedalmond = RGB(255, 235, 205)
+    val blueviolet = RGB(138, 43, 226)
+    val brown = RGB(165, 42, 42)
+    val burlywood = RGB(222, 184, 135)
+    val cadetblue = RGB(95, 158, 160)
+    val chartreuse = RGB(127, 255, 0)
+    val chocolate = RGB(210, 105, 30)
+    val coral = RGB(255, 127, 80)
+    val cornflowerblue = RGB(100, 149, 237)
+    val cornsilk = RGB(255, 248, 220)
+    val crimson = RGB(220, 20, 60)
+    val cyan = RGB(0, 255, 255)
+    val darkblue = RGB(0, 0, 139)
+    val darkcyan = RGB(0, 139, 139)
+    val darkgoldenrod = RGB(184, 134, 11)
+    val darkgray = RGB(169, 169, 169)
+    val darkgreen = RGB(0, 100, 0)
+    val darkgrey = RGB(169, 169, 169)
+    val darkkhaki = RGB(189, 183, 107)
+    val darkmagenta = RGB(139, 0, 139)
+    val darkolivegreen = RGB(85, 107, 47)
+    val darkorange = RGB(255, 140, 0)
+    val darkorchid = RGB(153, 50, 204)
+    val darkred = RGB(139, 0, 0)
+    val darksalmon = RGB(233, 150, 122)
+    val darkseagreen = RGB(143, 188, 143)
+    val darkslateblue = RGB(72, 61, 139)
+    val darkslategray = RGB(47, 79, 79)
+    val darkslategrey = RGB(47, 79, 79)
+    val darkturquoise = RGB(0, 206, 209)
+    val darkviolet = RGB(148, 0, 211)
+    val deeppink = RGB(255, 20, 147)
+    val deepskyblue = RGB(0, 191, 255)
+    val dimgray = RGB(105, 105, 105)
+    val dimgrey = RGB(105, 105, 105)
+    val dodgerblue = RGB(30, 144, 255)
+    val firebrick = RGB(178, 34, 34)
+    val floralwhite = RGB(255, 250, 240)
+    val forestgreen = RGB(34, 139, 34)
+    val gainsboro = RGB(220, 220, 220)
+    val ghostwhite = RGB(248, 248, 255)
+    val gold = RGB(255, 215, 0)
+    val goldenrod = RGB(218, 165, 32)
+    val greenyellow = RGB(173, 255, 47)
+    val grey = RGB(128, 128, 128)
+    val honeydew = RGB(240, 255, 240)
+    val hotpink = RGB(255, 105, 180)
+    val indianred = RGB(205, 92, 92)
+    val indigo = RGB(75, 0, 130)
+    val ivory = RGB(255, 255, 240)
+    val khaki = RGB(240, 230, 140)
+    val lavender = RGB(230, 230, 250)
+    val lavenderblush = RGB(255, 240, 245)
+    val lawngreen = RGB(124, 252, 0)
+    val lemonchiffon = RGB(255, 250, 205)
+    val lightblue = RGB(173, 216, 230)
+    val lightcoral = RGB(240, 128, 128)
+    val lightcyan = RGB(224, 255, 255)
+    val lightgoldenrodyellow = RGB(250, 250, 210)
+    val lightgray = RGB(211, 211, 211)
+    val lightgreen = RGB(144, 238, 144)
+    val lightgrey = RGB(211, 211, 211)
+    val lightpink = RGB(255, 182, 193)
+    val lightsalmon = RGB(255, 160, 122)
+    val lightseagreen = RGB(32, 178, 170)
+    val lightskyblue = RGB(135, 206, 250)
+    val lightslategray = RGB(119, 136, 153)
+    val lightslategrey = RGB(119, 136, 153)
+    val lightsteelblue = RGB(176, 196, 222)
+    val lightyellow = RGB(255, 255, 224)
+    val limegreen = RGB(50, 205, 50)
+    val linen = RGB(250, 240, 230)
+    val magenta = RGB(255, 0, 255)
+    val mediumaquamarine = RGB(102, 205, 170)
+    val mediumblue = RGB(0, 0, 205)
+    val mediumorchid = RGB(186, 85, 211)
+    val mediumpurple = RGB(147, 112, 219)
+    val mediumseagreen = RGB(60, 179, 113)
+    val mediumslateblue = RGB(123, 104, 238)
+    val mediumspringgreen = RGB(0, 250, 154)
+    val mediumturquoise = RGB(72, 209, 204)
+    val mediumvioletred = RGB(199, 21, 133)
+    val midnightblue = RGB(25, 25, 112)
+    val mintcream = RGB(245, 255, 250)
+    val mistyrose = RGB(255, 228, 225)
+    val moccasin = RGB(255, 228, 181)
+    val navajowhite = RGB(255, 222, 173)
+    val oldlace = RGB(253, 245, 230)
+    val olivedrab = RGB(107, 142, 35)
+    val orangered = RGB(255, 69, 0)
+    val orchid = RGB(218, 112, 214)
+    val palegoldenrod = RGB(238, 232, 170)
+    val palegreen = RGB(152, 251, 152)
+    val paleturquoise = RGB(175, 238, 238)
+    val palevioletred = RGB(219, 112, 147)
+    val papayawhip = RGB(255, 239, 213)
+    val peachpuff = RGB(255, 218, 185)
+    val peru = RGB(205, 133, 63)
+    val pink = RGB(255, 192, 203)
+    val plum = RGB(221, 160, 221)
+    val powderblue = RGB(176, 224, 230)
+    val rosybrown = RGB(188, 143, 143)
+    val royalblue = RGB(65, 105, 225)
+    val saddlebrown = RGB(139, 69, 19)
+    val salmon = RGB(250, 128, 114)
+    val sandybrown = RGB(244, 164, 96)
+    val seagreen = RGB(46, 139, 87)
+    val seashell = RGB(255, 245, 238)
+    val sienna = RGB(160, 82, 45)
+    val skyblue = RGB(135, 206, 235)
+    val slateblue = RGB(106, 90, 205)
+    val slategray = RGB(112, 128, 144)
+    val slategrey = RGB(112, 128, 144)
+    val snow = RGB(255, 250, 250)
+    val springgreen = RGB(0, 255, 127)
+    val steelblue = RGB(70, 130, 180)
+    val tan = RGB(210, 180, 140)
+    val thistle = RGB(216, 191, 216)
+    val tomato = RGB(255, 99, 71)
+    val turquoise = RGB(64, 224, 208)
+    val violet = RGB(238, 130, 238)
+    val wheat = RGB(245, 222, 179)
+    val whitesmoke = RGB(245, 245, 245)
+    val yellowgreen = RGB(154, 205, 50)
+    val rebeccapurple = RGB(102, 51, 153)
+    val transparent = RGB(0, 0, 0, 0f)
+
+    /**
+     * A map of CSS color names to RGB values.
+     */
+    val colorsByName = mapOf(
+            "black" to black,
+            "silver" to silver,
+            "gray" to gray,
+            "white" to white,
+            "maroon" to maroon,
+            "red" to red,
+            "purple" to purple,
+            "fuchsia" to fuchsia,
+            "green" to green,
+            "lime" to lime,
+            "olive" to olive,
+            "yellow" to yellow,
+            "navy" to navy,
+            "blue" to blue,
+            "teal" to teal,
+            "aqua" to aqua,
+            "orange" to orange,
+            "aliceblue" to aliceblue,
+            "antiquewhite" to antiquewhite,
+            "aquamarine" to aquamarine,
+            "azure" to azure,
+            "beige" to beige,
+            "bisque" to bisque,
+            "blanchedalmond" to blanchedalmond,
+            "blueviolet" to blueviolet,
+            "brown" to brown,
+            "burlywood" to burlywood,
+            "cadetblue" to cadetblue,
+            "chartreuse" to chartreuse,
+            "chocolate" to chocolate,
+            "coral" to coral,
+            "cornflowerblue" to cornflowerblue,
+            "cornsilk" to cornsilk,
+            "crimson" to crimson,
+            "cyan" to cyan,
+            "darkblue" to darkblue,
+            "darkcyan" to darkcyan,
+            "darkgoldenrod" to darkgoldenrod,
+            "darkgray" to darkgray,
+            "darkgreen" to darkgreen,
+            "darkgrey" to darkgrey,
+            "darkkhaki" to darkkhaki,
+            "darkmagenta" to darkmagenta,
+            "darkolivegreen" to darkolivegreen,
+            "darkorange" to darkorange,
+            "darkorchid" to darkorchid,
+            "darkred" to darkred,
+            "darksalmon" to darksalmon,
+            "darkseagreen" to darkseagreen,
+            "darkslateblue" to darkslateblue,
+            "darkslategray" to darkslategray,
+            "darkslategrey" to darkslategrey,
+            "darkturquoise" to darkturquoise,
+            "darkviolet" to darkviolet,
+            "deeppink" to deeppink,
+            "deepskyblue" to deepskyblue,
+            "dimgray" to dimgray,
+            "dimgrey" to dimgrey,
+            "dodgerblue" to dodgerblue,
+            "firebrick" to firebrick,
+            "floralwhite" to floralwhite,
+            "forestgreen" to forestgreen,
+            "gainsboro" to gainsboro,
+            "ghostwhite" to ghostwhite,
+            "gold" to gold,
+            "goldenrod" to goldenrod,
+            "greenyellow" to greenyellow,
+            "grey" to grey,
+            "honeydew" to honeydew,
+            "hotpink" to hotpink,
+            "indianred" to indianred,
+            "indigo" to indigo,
+            "ivory" to ivory,
+            "khaki" to khaki,
+            "lavender" to lavender,
+            "lavenderblush" to lavenderblush,
+            "lawngreen" to lawngreen,
+            "lemonchiffon" to lemonchiffon,
+            "lightblue" to lightblue,
+            "lightcoral" to lightcoral,
+            "lightcyan" to lightcyan,
+            "lightgoldenrodyellow" to lightgoldenrodyellow,
+            "lightgray" to lightgray,
+            "lightgreen" to lightgreen,
+            "lightgrey" to lightgrey,
+            "lightpink" to lightpink,
+            "lightsalmon" to lightsalmon,
+            "lightseagreen" to lightseagreen,
+            "lightskyblue" to lightskyblue,
+            "lightslategray" to lightslategray,
+            "lightslategrey" to lightslategrey,
+            "lightsteelblue" to lightsteelblue,
+            "lightyellow" to lightyellow,
+            "limegreen" to limegreen,
+            "linen" to linen,
+            "magenta" to magenta,
+            "mediumaquamarine" to mediumaquamarine,
+            "mediumblue" to mediumblue,
+            "mediumorchid" to mediumorchid,
+            "mediumpurple" to mediumpurple,
+            "mediumseagreen" to mediumseagreen,
+            "mediumslateblue" to mediumslateblue,
+            "mediumspringgreen" to mediumspringgreen,
+            "mediumturquoise" to mediumturquoise,
+            "mediumvioletred" to mediumvioletred,
+            "midnightblue" to midnightblue,
+            "mintcream" to mintcream,
+            "mistyrose" to mistyrose,
+            "moccasin" to moccasin,
+            "navajowhite" to navajowhite,
+            "oldlace" to oldlace,
+            "olivedrab" to olivedrab,
+            "orangered" to orangered,
+            "orchid" to orchid,
+            "palegoldenrod" to palegoldenrod,
+            "palegreen" to palegreen,
+            "paleturquoise" to paleturquoise,
+            "palevioletred" to palevioletred,
+            "papayawhip" to papayawhip,
+            "peachpuff" to peachpuff,
+            "peru" to peru,
+            "pink" to pink,
+            "plum" to plum,
+            "powderblue" to powderblue,
+            "rosybrown" to rosybrown,
+            "royalblue" to royalblue,
+            "saddlebrown" to saddlebrown,
+            "salmon" to salmon,
+            "sandybrown" to sandybrown,
+            "seagreen" to seagreen,
+            "seashell" to seashell,
+            "sienna" to sienna,
+            "skyblue" to skyblue,
+            "slateblue" to slateblue,
+            "slategray" to slategray,
+            "slategrey" to slategrey,
+            "snow" to snow,
+            "springgreen" to springgreen,
+            "steelblue" to steelblue,
+            "tan" to tan,
+            "thistle" to thistle,
+            "tomato" to tomato,
+            "turquoise" to turquoise,
+            "violet" to violet,
+            "wheat" to wheat,
+            "whitesmoke" to whitesmoke,
+            "yellowgreen" to yellowgreen,
+            "rebeccapurple" to rebeccapurple,
+            "transparent" to transparent
+    )
+}